Move QueueAuth inside pasplugin module
